WiktionaryParser behaviour (wiktionary_parser v0.1.0)

This module provides functions for parsing Russian Wiktionary entries given the dictionary form of a word.

Currently, only parsing of verbs and nouns is supported.

Callbacks

parse(word)

@callback parse(word :: String.t()) ::
  {:ok, WiktionaryParser.Verb.t() | WiktionaryParser.Noun.t()}
  | {:error, String.t()}

Parses the Wiktionary entry for a word of a particular part of speech given its dictionary form.

Returns {:ok, struct} if successful, {:error, reason} otherwise.

Functions

parse(word, part_of_speech)

@spec parse(word :: String.t(), part_of_speech :: :verb | :noun) ::
  {:ok, WiktionaryParser.Verb.t() | WiktionaryParser.Noun.t()}
  | {:error, String.t()}

Parses the Wiktionary entry for a word as the appropriate struct given its dictionary form and the part of speech.

The provided part_of_speech atom must be either :verb or :noun. Returns {:ok, struct} if successful, {:error, reason} otherwise.

Examples

iex> WiktionaryParser.parse("смотреть", :verb)
{:ok, %WiktionaryParser.Verb{translation: "to look", infinitive: "смотре́ть", ...}}

iex> WiktionaryParser.parse("кот", :noun)
{:ok, %WiktionaryParser.Noun{translation: "tomcat", gender: :masculine, ...}}

parse!(word, part_of_speech)

@spec parse!(word :: String.t(), part_of_speech :: :verb | :noun) ::
  WiktionaryParser.Verb.t() | WiktionaryParser.Noun.t()

Parses the Wiktionary entry for a word as the appropriate struct given its dictionary form and the part of speech.

The provided part_of_speech atom must be either :verb or :noun. Returns the parsed struct if successful, or raises an error otherwise.

Examples

iex> WiktionaryParser.parse!("смотреть", :verb)
%WiktionaryParser.Verb{translation: "to look", infinitive: "смотре́ть", ...}

iex> WiktionaryParser.parse!("кот", :noun)
%WiktionaryParser.Noun{translation: "tomcat", gender: :masculine, ...}